Title Screen Art
I added the title screen art to the title scene as well as some art for the interactable buttons. I also put in a place for a quit game button if we want.

You can also find all art in the images folder. The platforms that are the right size are still in prefabs.